backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.38
python-dotenv==1.0.1
requests==2.32.3
shapely==2.1.2
tenacity==9.0.0